+++ tests/actor_test.py
@@ -4,14 +4,8 @@ import uuid
 import time
 
 from pykka.actor import ThreadingActor
+from pykka.exceptions import ActorDeadError
 from pykka.registry import ActorRegistry
-
-try:
-    import gevent.event
-    from pykka.gevent import GeventActor
-    HAS_GEVENT = True
-except ImportError:
-    HAS_GEVENT = False
 
 
 class AnActor(object):
@@ -48,6 +42,8 @@ class AnActor(object):
             self.stop()
         elif message.get('command') == 'greetings':
             self.greetings_was_received.set()
+        elif message.get('command') == 'callback':
+            message['callback']()
         else:
             super(AnActor, self).on_receive(message)
 
@@ -130,6 +126,24 @@ class ActorTest(object):
     def tearDown(self):
         ActorRegistry.stop_all()
 
+    def test_messages_left_in_queue_after_actor_stops_receive_an_error(self):
+        event = self.event_class()
+        self.actor_ref.tell({'command': 'callback', 'callback': event.wait})
+        self.actor_ref.stop(block=False)
+        response = self.actor_ref.ask({'command': 'irrelevant'}, block=False)
+        event.set()
+
+        self.assertRaises(ActorDeadError, response.get, timeout=0.5)
+
+    def test_stop_requests_left_in_queue_after_actor_stops_are_handled(self):
+        event = self.event_class()
+        self.actor_ref.tell({'command': 'callback', 'callback': event.wait})
+        self.actor_ref.stop(block=False)
+        response = self.actor_ref.ask({'command': 'pykka_stop'}, block=False)
+        event.set()
+
+        response.get(timeout=0.5)
+
     def test_actor_has_an_uuid4_based_urn(self):
         self.assertEqual(4, uuid.UUID(self.actor_ref.actor_urn).version)
 
@@ -202,10 +216,10 @@ class ActorTest(object):
     def test_on_start_failure_causes_actor_to_stop(self):
         # Actor should not be alive if on_start fails.
         start_event = self.event_class()
-        actor = self.EarlyFailingActor.start(start_event)
+        actor_ref = self.EarlyFailingActor.start(start_event)
         start_event.wait(5)
-        time.sleep(0.01)
-        self.assertFalse(actor.is_alive())
+        actor_ref.actor_stopped.wait(5)
+        self.assertFalse(actor_ref.is_alive())
 
     def test_on_stop_is_called_when_actor_is_stopped(self):
         self.assertFalse(self.on_stop_was_called.is_set())
@@ -277,27 +291,32 @@ class ActorTest(object):
         self.assertEqual(0, len(ActorRegistry.get_all()))
 
 
-class ThreadingActorTest(ActorTest, unittest.TestCase):
-    event_class = threading.Event
+def ConcreteActorTest(actor_class, event_class):
+    class C(ActorTest, unittest.TestCase):
+        class AnActor(AnActor, actor_class):
+            pass
 
-    class AnActor(AnActor, ThreadingActor):
-        pass
+        class EarlyStoppingActor(EarlyStoppingActor, actor_class):
+            pass
 
-    class EarlyStoppingActor(EarlyStoppingActor, ThreadingActor):
-        pass
+        class EarlyFailingActor(EarlyFailingActor, actor_class):
+            pass
 
-    class EarlyFailingActor(EarlyFailingActor, ThreadingActor):
-        pass
+        class LateFailingActor(LateFailingActor, actor_class):
+            pass
 
-    class LateFailingActor(LateFailingActor, ThreadingActor):
-        pass
+        class FailingOnFailureActor(FailingOnFailureActor, actor_class):
+            pass
 
-    class FailingOnFailureActor(FailingOnFailureActor, ThreadingActor):
-        pass
+        class SuperInitActor(actor_class):
+            pass
 
-    class SuperInitActor(ThreadingActor):
-        pass
+    C.__name__ = '%sTest' % actor_class.__name__
+    C.event_class = event_class
+    return C
 
+
+class ThreadingActorTest(ConcreteActorTest(ThreadingActor, threading.Event)):
     class DaemonActor(ThreadingActor):
         use_daemon_thread = True
 
@@ -325,24 +344,16 @@ class ThreadingActorTest(ActorTest, unittest.TestCase):
         actor_ref.stop()
 
 
-if HAS_GEVENT:
-    class GeventActorTest(ActorTest, unittest.TestCase):
-        event_class = gevent.event.Event
+try:
+    import gevent.event
+    from pykka.gevent import GeventActor
+    GeventActorTest = ConcreteActorTest(GeventActor, gevent.event.Event)
+except ImportError:
+    pass
 
-        class AnActor(AnActor, GeventActor):
-            pass
-
-        class EarlyStoppingActor(EarlyStoppingActor, GeventActor):
-            pass
-
-        class EarlyFailingActor(EarlyFailingActor, GeventActor):
-            pass
-
-        class LateFailingActor(LateFailingActor, GeventActor):
-            pass
-
-        class FailingOnFailureActor(FailingOnFailureActor, GeventActor):
-            pass
-
-        class SuperInitActor(GeventActor):
-            pass
+try:
+    import eventlet  # noqa
+    from pykka.eventlet import EventletActor, EventletEvent
+    EventletActorTest = ConcreteActorTest(EventletActor, EventletEvent)
+except ImportError:
+    pass