+++ tests/logging_test.py
@@ -1,17 +1,9 @@
 import logging
 import threading
-import time
 import unittest
 
 from pykka.actor import ThreadingActor
 from pykka.registry import ActorRegistry
-
-try:
-    import gevent.event
-    from pykka.gevent import GeventActor
-    HAS_GEVENT = True
-except ImportError:
-    HAS_GEVENT = False
 
 from tests import TestLogHandler
 from tests.actor_test import (
@@ -44,22 +36,24 @@ class ActorLoggingTest(object):
 
     def test_unexpected_messages_are_logged(self):
         self.actor_ref.ask({'unhandled': 'message'})
-        self.assertEqual(1, len(self.log_handler.messages['warning']))
-        log_record = self.log_handler.messages['warning'][0]
+        self.log_handler.wait_for_message('warning')
+        with self.log_handler.lock:
+            self.assertEqual(1, len(self.log_handler.messages['warning']))
+            log_record = self.log_handler.messages['warning'][0]
         self.assertEqual(
             'Unexpected message received by %s' % self.actor_ref,
             log_record.getMessage().split(': ')[0])
 
     def test_exception_is_logged_when_returned_to_caller(self):
-        # pylint: disable = W0703
         try:
             self.actor_proxy.raise_exception().get()
             self.fail('Should raise exception')
         except Exception:
             pass
-        # pylint: enable = W0703
-        self.assertEqual(1, len(self.log_handler.messages['debug']))
-        log_record = self.log_handler.messages['debug'][0]
+        self.log_handler.wait_for_message('debug')
+        with self.log_handler.lock:
+            self.assertEqual(1, len(self.log_handler.messages['debug']))
+            log_record = self.log_handler.messages['debug'][0]
         self.assertEqual(
             'Exception returned from %s to caller:' % self.actor_ref,
             log_record.getMessage())
@@ -71,8 +65,10 @@ class ActorLoggingTest(object):
         self.actor_ref.tell({'command': 'raise exception'})
         self.on_failure_was_called.wait(5)
         self.assertTrue(self.on_failure_was_called.is_set())
-        self.assertEqual(1, len(self.log_handler.messages['error']))
-        log_record = self.log_handler.messages['error'][0]
+        self.log_handler.wait_for_message('error')
+        with self.log_handler.lock:
+            self.assertEqual(1, len(self.log_handler.messages['error']))
+            log_record = self.log_handler.messages['error'][0]
         self.assertEqual(
             'Unhandled exception in %s:' % self.actor_ref,
             log_record.getMessage())
@@ -85,8 +81,10 @@ class ActorLoggingTest(object):
         self.actor_ref.tell({'command': 'raise base exception'})
         self.on_stop_was_called.wait(5)
         self.assertTrue(self.on_stop_was_called.is_set())
-        self.assertEqual(3, len(self.log_handler.messages['debug']))
-        log_record = self.log_handler.messages['debug'][0]
+        self.log_handler.wait_for_message('debug', num_messages=3)
+        with self.log_handler.lock:
+            self.assertEqual(3, len(self.log_handler.messages['debug']))
+            log_record = self.log_handler.messages['debug'][0]
         self.assertEqual(
             'BaseException() in %s. Stopping all actors.' % self.actor_ref,
             log_record.getMessage())
@@ -96,9 +94,10 @@ class ActorLoggingTest(object):
         start_event = self.event_class()
         actor_ref = self.EarlyFailingActor.start(start_event)
         start_event.wait(5)
-        time.sleep(0.01)  # Too ensure that the log handler is updated
-        self.assertEqual(1, len(self.log_handler.messages['error']))
-        log_record = self.log_handler.messages['error'][0]
+        self.log_handler.wait_for_message('error')
+        with self.log_handler.lock:
+            self.assertEqual(1, len(self.log_handler.messages['error']))
+            log_record = self.log_handler.messages['error'][0]
         self.assertEqual(
             'Unhandled exception in %s:' % actor_ref,
             log_record.getMessage())
@@ -108,9 +107,10 @@ class ActorLoggingTest(object):
         stop_event = self.event_class()
         actor_ref = self.LateFailingActor.start(stop_event)
         stop_event.wait(5)
-        time.sleep(0.01)  # Too ensure that the log handler is updated
-        self.assertEqual(1, len(self.log_handler.messages['error']))
-        log_record = self.log_handler.messages['error'][0]
+        self.log_handler.wait_for_message('error')
+        with self.log_handler.lock:
+            self.assertEqual(1, len(self.log_handler.messages['error']))
+            log_record = self.log_handler.messages['error'][0]
         self.assertEqual(
             'Unhandled exception in %s:' % actor_ref,
             log_record.getMessage())
@@ -121,9 +121,10 @@ class ActorLoggingTest(object):
         actor_ref = self.FailingOnFailureActor.start(failure_event)
         actor_ref.tell({'command': 'raise exception'})
         failure_event.wait(5)
-        time.sleep(0.01)  # Too ensure that the log handler is updated
-        self.assertEqual(2, len(self.log_handler.messages['error']))
-        log_record = self.log_handler.messages['error'][0]
+        self.log_handler.wait_for_message('error', num_messages=2)
+        with self.log_handler.lock:
+            self.assertEqual(2, len(self.log_handler.messages['error']))
+            log_record = self.log_handler.messages['error'][0]
         self.assertEqual(
             'Unhandled exception in %s:' % actor_ref,
             log_record.getMessage())
@@ -153,34 +154,43 @@ class AnActor(object):
         raise Exception('foo')
 
 
-class ThreadingActorLoggingTest(ActorLoggingTest, unittest.TestCase):
-    event_class = threading.Event
-
-    class AnActor(AnActor, ThreadingActor):
-        pass
-
-    class EarlyFailingActor(EarlyFailingActor, ThreadingActor):
-        pass
-
-    class LateFailingActor(LateFailingActor, ThreadingActor):
-        pass
-
-    class FailingOnFailureActor(FailingOnFailureActor, ThreadingActor):
-        pass
-
-
-if HAS_GEVENT:
-    class GeventActorLoggingTest(ActorLoggingTest, unittest.TestCase):
-        event_class = gevent.event.Event
-
-        class AnActor(AnActor, GeventActor):
+def ConcreteActorLoggingTest(actor_class, event_class):
+    class C(ActorLoggingTest, unittest.TestCase):
+        class AnActor(AnActor, actor_class):
             pass
 
-        class EarlyFailingActor(EarlyFailingActor, GeventActor):
+        class EarlyFailingActor(EarlyFailingActor, actor_class):
             pass
 
-        class LateFailingActor(LateFailingActor, GeventActor):
+        class LateFailingActor(LateFailingActor, actor_class):
             pass
 
-        class FailingOnFailureActor(FailingOnFailureActor, GeventActor):
+        class FailingOnFailureActor(FailingOnFailureActor, actor_class):
             pass
+
+    C.event_class = event_class
+    C.__name__ = '%sLoggingTest' % actor_class.__name__
+    return C
+
+
+ThreadingActorLoggingTest = ConcreteActorLoggingTest(
+    ThreadingActor, threading.Event)
+
+
+try:
+    import gevent.event
+    from pykka.gevent import GeventActor
+
+    GeventActorLoggingTest = ConcreteActorLoggingTest(
+        GeventActor, gevent.event.Event)
+except ImportError:
+    pass
+
+
+try:
+    from pykka.eventlet import EventletActor, EventletEvent
+
+    EventletActorLoggingTest = ConcreteActorLoggingTest(
+        EventletActor, EventletEvent)
+except ImportError:
+    pass